Switch to Burst Compiler for animation to use batching
I installed the Burst and Collections packages to enable batching for 2D animation. The speedup was necessary for the game to even run. Currently most of the processing time seems to be taken up by the editor. One troubling thing to note is that the editor has crashed several times after installing the burst compiler.